Module Catapult_client

Backend for Catapult, using a connection to the daemon.

module Endpoint_address = Catapult_utils.Endpoint_address

module Backend : sig ... end

module Connection : sig ... end

val default_endpoint : Endpoint_address.t

val with_conn : addr:Catapult_utils.Endpoint_address.t -> ?trace_id:string -> unit -> (Connection.t -> 'a) -> 'a

val backend_of_conn : Connection.t -> Catapult.backend

val addr_of_string_exn : string -> Endpoint_address.t

Parse a remote address.

val trace_collector_of_conn : Connection.t -> Trace_core.collector

Obtain a trace collector from a network connection

val with_ : addr:Catapult_utils.Endpoint_address.t -> ?trace_id:string -> unit -> (unit -> 'a) -> 'a

with_ ~addr () f runs f() in an environment where a connection to addr has been established and is used to forward tracing events to the remote daemon.